tiendav4/requirements.txt
django-crispy-forms
django-autoslug
Pillow